+++ CHANGELOG.md
@@ -10,7 +10,9 @@ Version 1.4.2
 - add 'rf' protocol handler (URLs has format: rf://IP:PORT)
 - add setup program
 - log error when RFA cannot be loaded
-- improve performance of rendering to texture (e.g. mirrors and scanners)
+- improve render to texture performance (e.g. mirrors and scanners)
+- improve 2D drawing performance
+- improve level loading time
 - fix sending crash reports
 - fix non-working face scrolling before first geomod use in custom levels (RF bug)
 - fix memory and VRAM leak when quick saving (RF bug)